reorganize file for current feature
