+++ package.json
@@ -86,6 +86,7 @@
     "@typescript-eslint/parser": "^5.7.0",
     "@ui-kitten/metro-config": "^5.1.2",
     "babel-jest": "^26.6.3",
+    "babel-plugin-module-resolver": "^5.0.0",
     "babel-plugin-transform-remove-console": "^6.9.4",
     "eslint": "^7.14.0",
     "jest": "^26.6.3",
